convert to multi-steps
