+++ ffmpeg-processing/src/workflows/audio-convert/index.ts
@@ -1,14 +1,25 @@
 import type { StreamingAudioInput } from "../../../types.js";
 import {
-	normalizeMetadataStep,
-	transcodeAndCompressStep,
+  normalizeMetadataStep,
+  createSandbox,
+  setupFfmpeg,
+  transcode,
+  streamOutput,
+  stopSandbox,
 } from "./steps/index.js";
 
 export async function compressAudioWorkflow(input: StreamingAudioInput) {
-	"use workflow";
+  "use workflow";
 
-	const normalized = await normalizeMetadataStep(input);
-	const metadata = await transcodeAndCompressStep(normalized);
+  const normalized = await normalizeMetadataStep(input);
+  const sandboxId = await createSandbox();
 
-	return metadata;
+  try {
+    await setupFfmpeg(sandboxId);
+    await transcode(normalized, sandboxId);
+    const metadata = await streamOutput(sandboxId, normalized.metadata.filename);
+    return metadata;
+  } finally {
+    await stopSandbox(sandboxId);
+  }
 }
