+++ ffmpeg-processing/src/workflows/audio-convert/steps/index.ts
@@ -1,2 +1,6 @@
 export { normalizeMetadataStep } from "./normalize-metadata.js";
-export { transcodeAndCompressStep } from "./transcode-and-compress.js";
+export { createSandbox } from "./create-sandbox.js";
+export { setupFfmpeg } from "./setup-ffmpeg.js";
+export { transcode } from "./transcode.js";
+export { streamOutput } from "./stream-output.js";
+export { stopSandbox } from "./stop-sandbox.js";
