+++ ffmpeg-processing/src/workflows/audio-convert/steps/stop-sandbox.ts
@@ -1,0 +1,9 @@
+import { Sandbox } from "@vercel/sandbox";
+
+export async function stopSandbox(sandboxId: string): Promise<void> {
+  "use step";
+
+  console.log("[Sandbox] Stopping...");
+  const sandbox = await Sandbox.get({ sandboxId });
+  await sandbox.stop().catch(() => {});
+}
